Version 1.9
  * More BSD services: utx, apm, apmd, auditdistd, ctld, ftp-proxy, kfd, swapon, pwcheck, mrouted, rpcstatd, rpclockd, keyserv, iscsid, iscsictl, sysctl, yppasswdd, ypxfrd, ypupdated, ypset, ypserv, pppoed, msgs, vixiecron
  * More Linux services: keyboard-setup, console-setup, pppd-dns, kerneloops, acpid
  * A new rcctl shim supplies limited compatibility with the OpenBSD tool.
  * Added convert-rcconf-presets for converting /etc/rc.conf{,.local} .
  * Added a vc-ncurses-io command.
  * The motd service is now named motd-dynamic, per the Debian init.d script of that name, allowing a BSD service to be named motd.

+++ source/cyclog.cpp
@@ -346,7 +346,7 @@ cyclog (
 		margin = m;
 	} catch (const popt::error & e) {
 		std::fprintf(stderr, "%s: FATAL: %s: %s\n", prog, e.arg, e.msg);
-		throw EXIT_USAGE;
+		throw static_cast<int>(EXIT_USAGE);
 	}
 	if (args.empty()) {
 		std::fprintf(stderr, "%s: FATAL: %s\n", prog, "Missing log directory name.");
